+++ examples/react/players/FullControl.js
@@ -11,12 +11,14 @@ class AutoPlay extends React.Component {
       playing: false,
       loaded: false,
       loop: false,
-      mute: false
+      mute: false,
+      volume: 1.0
     }
     this.handleToggle = this.handleToggle.bind(this)
     this.handleOnLoad = this.handleOnLoad.bind(this)
     this.handleOnEnd = this.handleOnEnd.bind(this)
     this.handleOnPlay = this.handleOnPlay.bind(this)
+    this.handleStop = this.handleStop.bind(this)
     this.renderSeekPos = this.renderSeekPos.bind(this)
     this.handleLoopToggle = this.handleLoopToggle.bind(this)
     this.handleMuteToggle = this.handleMuteToggle.bind(this)
@@ -51,6 +53,14 @@ class AutoPlay extends React.Component {
       playing: false
     })
     this.clearRAF()
+  }
+
+  handleStop () {
+    this.player.stop()
+    this.setState({
+      playing: false // Need to update our local state so we don't immediately invoke autoplay
+    })
+    this.renderSeekPos()
   }
 
   handleLoopToggle () {
@@ -89,10 +99,14 @@ class AutoPlay extends React.Component {
           onEnd={this.handleOnEnd}
           loop={this.state.loop}
           mute={this.state.mute}
+          volume={this.state.volume}
           ref={(ref) => this.player = ref}
         />
         <button onClick={this.handleToggle}>
           {(this.state.playing) ? 'Pause' : 'Play' }
+        </button>
+        <button onClick={this.handleStop}>
+          Stop
         </button>
         <p>{(this.state.loaded) ? 'Loaded' : 'Loading'}</p>
         <p>
@@ -117,6 +131,20 @@ class AutoPlay extends React.Component {
               onChange={this.handleMuteToggle}
             />
         </label>
+        <div>
+          <label>
+            <input
+              type='range'
+              min='0'
+              max='1'
+              step='.05'
+              value={this.state.volume}
+              onChange={e => this.setState({volume: parseFloat(e.target.value)})}
+            />
+            <br /> Volume: {this.state.volume}
+          </label>
+        </div>
+
       </div>
     )
   }